Moved include/IWalletLegacy.h to lib/WalletLegacy/IWalletLegacy.h
