+++ lib/WalletLegacy/IWalletLegacy.h
